move components into the components library
