+++ frontend/app/src/components/UniversalSidebar.tsx
